Sepaated the stacks to enable CFN parameters
