+++ cdk/README.md
@@ -24,19 +24,19 @@ The scripts will update the Lambda function zip file or Lambda layer zip file re
 
 To prepare your environment for CDK usage, run 
 
-    cdk bootstrap -c elbTargetGroupArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:targetgroup/AppServerATG/090a4ba28ada9d48"
+    cdk bootstrap -c elbTargetGroupArn="REPLACE_WITH_YOUR_ELBTARGETGROUP_ARN"
 
 To synthesize the CloudFormation templates:
 
-    cdk synth -c elbTargetGroupArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:targetgroup/AppServerATG/090a4ba28ada9d48"
+    cdk synth -c elbTargetGroupArn="REPLACE_WITH_YOUR_ELBTARGETGROUP_ARN"
 
 Replace the *elbTargetGroupArn* context parameter value with the ARN for your desired Target Group. The stack uses the elbTargetGroupArn parameter value to generate the correct CF template. As a prerequisite, you must have Application Load Balancers with appropriate Target Groups in your AWS environment.
 
-To deploy the stacks:
+To deploy the ALBMonitorStack:
 
-    cdk deploy ALBMonitorStack -c elbTargetGroupArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:targetgroup/AppServerATG/090a4ba28ada9d48" --parameters elbArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:loadbalancer/app/AgentPortalALB/bb6bb42b08f94c0b" --parameters elbListenerArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:listener/app/AgentPortalALB/bb6bb42b08f94c0b/b3784a6b090b3696"
+    cdk deploy ALBMonitorStack -c elbTargetGroupArn="REPLACE_WITH_YOUR_ELBTARGETGROUP_ARN" --parameters elbArn="REPLACE_WITH_YOUR_ELB_ARN" --parameters elbListenerArn="REPLACE_WITH_YOUR_ELBLISTENER_ARN"
 
-    cdk deploy ALBCloudWatchStack -c elbTargetGroupArn="arn:aws:elasticloadbalancing:SOME_REGION:SOME_ACCOUNT_NUM:targetgroup/AppServerATG/090a4ba28ada9d48" --parameters cwAlarmThreshold=500
+Note: Deployment of ALBMonitorStack will output the ARN of the Lambda provide to ALBCloudWatchStack.
 
 Provide the appropriate parameter values. The follow parameters are required:
 
@@ -44,9 +44,24 @@ Provide the appropriate parameter values. The follow parameters are required:
     elbListenerARN - ARN for the ALB Listener to update for shedding/restoring
 
 The following parameters are optional:
+    elbShedPercent - Percentage to shed expressed as an integer. Default: 5
+    maxElbShedPercent - aximum allowable load to shed from ELB. Default: 100
+    elbRestorePercent - Percentage to restore expressed as an integer. Default: 5
+    shedMesgDelaySec - Number of seconds to delay shed messages. Default: 60
+    restoreMesgDelaySec - Number of seconds to delay restore messages. Default: 120
 
+To deploy the ALBCloudWatchStack:
+
+    cdk deploy ALBCloudWatchStack -c elbTargetGroupArn="REPLACE_WITH_YOUR_ELBTARGETGROUP_ARN" --parameters cwAlarmLambdaArn="REPLACE_WITH_YOUR_LAMBDA_ARN" --
+
+Provide the appropriate parameter values. The follow parameters are required:
+    cwAlarmLambdaArn - ARN for Lambda to fire when in Alarm state. 
+
+The following parameters are optional:
     cwAlarmNamespace - The namespace for the CloudWatch (CW) metric (https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/viewing_metrics_with_cloudwatch.html). Default: AWS/ApplicationELB
     cwAlarmMetricName - The name of the CW metric. Default: RequestCountPerTarget
-    cwAlarmMetricStatistic - Function to use for aggregating the statistic. Default: Average
+    cwAlarmMetricStat - Function to use for aggregating the statistic. Default: Average
     cwAlarmThreshold - Threshold value that triggers an alarm for the metric. Default: 500
     cwAlarmPeriods - The evaluation period for the alarm. Default: 3.
+
+
